Refactor classes in test folder to match main folder
