+++ src/test/java/seedu/address/logic/commands/quiz/QuizCommandTestUtil.java
@@ -1,8 +1,10 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.quiz;
 
 import static org.junit.Assert.assertEquals;
 
 import seedu.address.logic.CommandHistory;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.modelmanager.quiz.QuizModel;
