+++ src/test/java/seedu/address/model/modelmanager/management/ManagementModelManagerTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model;
+package seedu.address.model.modelmanager.management;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -15,8 +15,9 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 import seedu.address.commons.core.GuiSettings;
+import seedu.address.model.Lessons;
+import seedu.address.model.UserPrefs;
 import seedu.address.model.lesson.Lesson;
-import seedu.address.model.modelmanager.management.ManagementModelManager;
 
 public class ManagementModelManagerTest {
     @Rule
@@ -39,9 +40,9 @@ public class ManagementModelManagerTest {
 
     @Test
     public void constructor() {
-        assertEquals(new UserPrefs(), modelManager.getUserPrefs());
+        Assert.assertEquals(new UserPrefs(), modelManager.getUserPrefs());
         assertEquals(new GuiSettings(), modelManager.getGuiSettings());
-        assertEquals(new Lessons().getLessons(), modelManager.getLessons());
+        Assert.assertEquals(new Lessons().getLessons(), modelManager.getLessons());
     }
 
     @Test
