+++ src/test/java/seedu/address/model/modelmanager/quiz/QuizCardTest.java
@@ -1,4 +1,4 @@
-package seedu.address.quiz;
+package seedu.address.model.modelmanager.quiz;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -10,8 +10,6 @@ import java.util.List;
 
 import org.junit.Test;
 
-import seedu.address.model.modelmanager.quiz.Quiz;
-import seedu.address.model.modelmanager.quiz.QuizCard;
 import seedu.address.testutil.Assert;
 
 public class QuizCardTest {
